+++ packages/odata-service-inquirer/test/unit/prompts/connectionValidator.test.ts
@@ -1,5 +1,5 @@
 import * as axiosExtension from '@sap-ux/axios-extension';
-import type { ODataServiceInfo } from '@sap-ux/axios-extension';
+import type { AbapServiceProvider, ODataServiceInfo } from '@sap-ux/axios-extension';
 import { ODataService, ODataVersion, ServiceProvider, type AxiosRequestConfig } from '@sap-ux/axios-extension';
 import type { ServiceInfo } from '@sap-ux/btp-utils';
 import {
@@ -12,6 +12,8 @@ import { AxiosError, type AxiosResponse } from 'axios';
 import { ERROR_TYPE, ErrorHandler } from '@sap-ux/inquirer-common';
 import { initI18nOdataServiceInquirer, t } from '../../../src/i18n';
 import { ConnectionValidator } from '../../../src/prompts/connectionValidator';
+import type { ConnectedSystem } from '../../../src/types';
+import { BackendSystem } from '@sap-ux/store';
 
 const odataServicesMock: ODataServiceInfo[] = [];
 const catalogServiceMock = jest.fn().mockImplementation(() => ({
@@ -436,7 +438,7 @@ describe('ConnectionValidator', () => {
             },
             systemid: 'abap_btp_001'
         };
-        const connectValidator = new ConnectionValidator();
+        let connectValidator = new ConnectionValidator();
         expect(await connectValidator.validateServiceInfo(serviceInfoMock as ServiceInfo)).toBe(true);
         expect(createAbapOnCloudProviderSpy).toHaveBeenCalledWith(
             expect.objectContaining({
@@ -458,6 +460,8 @@ describe('ConnectionValidator', () => {
         (connectValidator.serviceProvider as any).refreshTokenChangedCb('newToken1234');
         expect(connectValidator.refreshToken).toEqual('newToken1234');
 
+        connectValidator = new ConnectionValidator();
+        createAbapOnCloudProviderSpy.mockClear();
         // Ensure refresh token is used to create a connection if presented
         expect(
             await connectValidator.validateServiceInfo(serviceInfoMock as ServiceInfo, undefined, '123refreshToken456')
@@ -470,6 +474,13 @@ describe('ConnectionValidator', () => {
                 refreshToken: '123refreshToken456'
             })
         );
+
+        createAbapOnCloudProviderSpy.mockClear();
+        // Should not create a new connection if the service url is the same as current valdidate url
+        expect(
+            await connectValidator.validateServiceInfo(serviceInfoMock as ServiceInfo, undefined, '123refreshToken456')
+        ).toBe(true);
+        expect(createAbapOnCloudProviderSpy).not.toHaveBeenCalled();
     });
 
     test('should attempt to validate auth using v4 catalog where v2 is not available or user is not authorized', async () => {
@@ -717,5 +728,76 @@ describe('ConnectionValidator', () => {
             errorType: ERROR_TYPE.AUTH,
             valResult: 'Authentication incorrect. Please check the SAP BTP destination authentication configuration.'
         });
+    });
+
+    test('should re-use `connectedSystem` when provided rather than re-authentication', async () => {
+        let connectValidator = new ConnectionValidator();
+        (connectValidator as any)._validatedUrl = 'https://system1:12345/';
+        const checkUrlSpy = jest.spyOn(connectValidator as any, 'checkUrl');
+        // Reentrance ticket, prevent re-authentication
+        let cachedConnectedSystem: ConnectedSystem = {
+            serviceProvider: {
+                catalog: catalogServiceMock
+            } as unknown as AbapServiceProvider,
+            backendSystem: {
+                name: 'system1',
+                url: 'https://system1:12345/',
+                authenticationType: 'reentranceTicket',
+                userDisplayName: 'user1',
+                client: '001'
+            }
+        };
+        connectValidator.setConnectedSystem(cachedConnectedSystem);
+        // Validate the connection has updated the connectionValidator properties correctly
+        expect(connectValidator.serviceProvider).toEqual(cachedConnectedSystem.serviceProvider);
+        expect(connectValidator.catalogs[ODataVersion.v2]).toBeDefined();
+        expect(connectValidator.catalogs[ODataVersion.v4]).toBeDefined();
+        expect(connectValidator.validatedUrl).toEqual(cachedConnectedSystem.backendSystem!.url);
+        expect(connectValidator.connectedUserName).toEqual(cachedConnectedSystem.backendSystem!.userDisplayName);
+        expect(connectValidator.validatedClient).toEqual(cachedConnectedSystem.backendSystem!.client);
+        expect(connectValidator.refreshToken).toEqual(undefined);
+        expect(connectValidator.validity).toEqual({
+            authenticated: true,
+            reachable: true,
+            urlFormat: true,
+            authRequired: true
+        });
+        expect(connectValidator.systemAuthType).toEqual('reentranceTicket');
+
+        let connectValResult = await connectValidator.validateUrl(cachedConnectedSystem.backendSystem!.url, {
+            isSystem: true,
+            systemAuthType: 'reentranceTicket'
+        });
+
+        expect(connectValResult).toEqual(true);
+        expect(checkUrlSpy).not.toHaveBeenCalled();
+
+        // Service Keys, prevent re-authentication
+        cachedConnectedSystem = {
+            serviceProvider: {
+                catalog: catalogServiceMock
+            } as unknown as AbapServiceProvider,
+            backendSystem: {
+                name: 'system2',
+                url: 'https://system2:1234554321/',
+                authenticationType: '',
+                serviceKeys: {
+                    url: 'https://system2:54321/'
+                }
+            }
+        };
+
+        connectValidator = new ConnectionValidator();
+        const createSystemConnectionSpy = jest.spyOn(connectValidator as any, 'checkUrl');
+        connectValidator.setConnectedSystem(cachedConnectedSystem);
+
+        connectValResult = await connectValidator.validateServiceInfo(
+            cachedConnectedSystem.backendSystem!.serviceKeys as ServiceInfo,
+            undefined,
+            'refreshToken1234'
+        );
+
+        expect(connectValResult).toEqual(true);
+        expect(createSystemConnectionSpy).not.toHaveBeenCalled();
     });
 });